requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
boto3